Add a function to create settings from templates
this allows the use of a template with many hocon settings to be used to
quickly define the things we care about without having to write 500
lines of Puppet code. This is desireable because Puppet code should be
clear and easy to understand, and 500 lines of individual resources is
not that.